Move CMake things to base conanfile
