+++ recipes/base_conanfile/config.yml
@@ -1,3 +1,3 @@
 versions:
-  v0.1.0:
+  v0.2.0:
     folder: all
