+++ recipes/composite/all/conanfile.py
@@ -5,9 +5,9 @@ from conans.errors import ConanInvalidConfiguration
 required_conan_version = ">=1.33.0"
 
 class CompositeConan(ConanFile):
-    python_requires = "base_conanfile/v0.1.0@jgsogo/stable"
-    python_requires_extend = "base_conanfile.BaseConanfile"
-
+    python_requires = "base_conanfile/v0.2.0@jgsogo/stable"
+    python_requires_extend = "base_conanfile.BaseCMakeConanfile"
+    
     name = "composite"
     homepage = "https://github.com/jgsogo/composite"
     description = "Data structures with composite nodes"
