+++ recipes/render_context/all/conanfile.py
@@ -1,12 +1,11 @@
 import os
-from conans import ConanFile, tools, CMake
-from conans.errors import ConanInvalidConfiguration
+from conans import ConanFile
 
 required_conan_version = ">=1.33.0"
 
 class RenderContextConan(ConanFile):
-    python_requires = "base_conanfile/v0.1.0@jgsogo/stable"
-    python_requires_extend = "base_conanfile.BaseConanfile"
+    python_requires = "base_conanfile/v0.2.0@jgsogo/stable"
+    python_requires_extend = "base_conanfile.BaseCMakeConanfile"
 
     name = "render_context"
     homepage = "https://github.com/jgsogo/render_context"
@@ -27,17 +26,6 @@ class RenderContextConan(ConanFile):
         self.requires("magnum/2020.06@jgsogo/stable")
         self.requires("magnum-integration/2020.06")
         self.requires("catch2/2.13.7")
-
-    def _configure_cmake(self):
-        if self._cmake:
-            return self._cmake
-        self._cmake = CMake(self)
-        self._cmake.configure()
-        return self._cmake
-
-    def build(self):
-        cmake = self._configure_cmake()
-        cmake.build()
 
     def package_info(self):
         self.cpp_info.defines.append("IMGUI_USER_CONFIG=\"{}\"".format(str(os.path.join(self.package_folder, "include", "render", "imgui", "imconfig.h"))))
